Move load_from_snowsql_config_file into connections.util
