+++ lib/streamlit/connections/util.py
@@ -13,7 +13,11 @@
 # limitations under the License.
 
 
+import configparser
+import os
 from typing import Any, Collection, Dict
+
+SNOWSQL_CONNECTION_FILE = "~/.snowsql/config"
 
 
 def extract_from_dict(
@@ -40,6 +44,33 @@ def extract_from_dict(
             d[k] = source_dict.pop(k)
 
     return d
+
+
+def load_from_snowsql_config_file(connection_name: str) -> Dict[str, Any]:
+    """Loads the dictionary from snowsql config file."""
+    snowsql_config_file = os.path.expanduser(SNOWSQL_CONNECTION_FILE)
+    if not os.path.exists(snowsql_config_file):
+        return {}
+
+    config = configparser.ConfigParser(inline_comment_prefixes="#")
+    config.read(snowsql_config_file)
+
+    if f"connections.{connection_name}" in config:
+        raw_conn_params = config[f"connections.{connection_name}"]
+    elif "connections" in config:
+        raw_conn_params = config["connections"]
+    else:
+        return {}
+
+    conn_params = {
+        k.replace("name", ""): v.strip('"') for k, v in raw_conn_params.items()
+    }
+
+    if "db" in conn_params:
+        conn_params["database"] = conn_params["db"]
+        del conn_params["db"]
+
+    return conn_params
 
 
 def running_in_sis() -> bool:
