+++ lib/tests/streamlit/connections/snowpark_connection_test.py
@@ -14,13 +14,12 @@
 
 import threading
 import unittest
-from unittest.mock import MagicMock, PropertyMock, mock_open, patch
+from unittest.mock import MagicMock, PropertyMock, patch
 
 import pytest
 
 import streamlit as st
 from streamlit.connections import SnowparkConnection
-from streamlit.connections.snowpark_connection import _load_from_snowsql_config_file
 from streamlit.errors import StreamlitAPIException
 from streamlit.runtime.scriptrunner import add_script_run_ctx
 from streamlit.runtime.secrets import AttrDict
@@ -31,53 +30,6 @@ from tests.testutil import create_mock_script_run_ctx
 class SnowparkConnectionTest(unittest.TestCase):
     def tearDown(self) -> None:
         st.cache_data.clear()
-
-    def test_load_from_snowsql_config_file_no_file(self):
-        assert _load_from_snowsql_config_file("my_snowpark_connection") == {}
-
-    @patch(
-        "streamlit.connections.snowpark_connection.os.path.exists",
-        MagicMock(return_value=True),
-    )
-    def test_load_from_snowsql_config_file_no_section(self):
-        with patch("builtins.open", new_callable=mock_open, read_data=""):
-            assert _load_from_snowsql_config_file("my_snowpark_connection") == {}
-
-    @patch(
-        "streamlit.connections.snowpark_connection.os.path.exists",
-        MagicMock(return_value=True),
-    )
-    def test_load_from_snowsql_config_file_named_section(self):
-        config_data = """
-[connections.my_snowpark_connection]
-accountname = "hello"
-dbname = notPostgres
-
-[connections]
-accountname = "i get overwritten"
-schemaname = public
-"""
-        with patch("builtins.open", new_callable=mock_open, read_data=config_data):
-            assert _load_from_snowsql_config_file("my_snowpark_connection") == {
-                "account": "hello",
-                "database": "notPostgres",
-            }
-
-    @patch(
-        "streamlit.connections.snowpark_connection.os.path.exists",
-        MagicMock(return_value=True),
-    )
-    def test_load_from_snowsql_config_file_default_section(self):
-        config_data = """
-[connections]
-accountname = "not overwritten"
-schemaname = public
-"""
-        with patch("builtins.open", new_callable=mock_open, read_data=config_data):
-            assert _load_from_snowsql_config_file("my_snowpark_connection") == {
-                "account": "not overwritten",
-                "schema": "public",
-            }
 
     @patch(
         "snowflake.snowpark.context.get_active_session",
@@ -92,7 +44,7 @@ schemaname = public
         assert conn._instance == "some active session"
 
     @patch(
-        "streamlit.connections.snowpark_connection._load_from_snowsql_config_file",
+        "streamlit.connections.snowpark_connection.load_from_snowsql_config_file",
         MagicMock(
             return_value={"account": "some_val_1", "password": "i get overwritten"}
         ),
